int64_t loop index

+++ src/ops/matmul/batched_gemm.h
@@ -5,7 +5,7 @@
 #include "gemm.h"
 
 template <typename T>
-T defaultEpilogue(float acc, size_t batch_idx, size_t row_idx, size_t col_idx) {
+T defaultEpilogue(float acc, int64_t batch_idx, int64_t row_idx, int64_t col_idx) {
     return llaisys::utils::cast<T>(acc);
 }
 
@@ -34,14 +34,14 @@ void batch_gemm_cpu_blocked_omp(
 
     omp_set_nested(1);
 #pragma omp parallel for schedule(static)
-  for (int ib = 0; ib < batch; ++ib) {
+  for (int64_t ib = 0; ib < (int64_t)batch; ++ib) {
     const T* A_batch = A + ib * batch_stride_A;
     const T* B_batch = B + ib * batch_stride_B;
     T*       C_batch = C + ib * batch_stride_C;
     gemm_cpu_blocked_omp<T, LayoutB>(A_batch, B_batch, C_batch,
                          M, N, K,
                          lda, ldb, ldc,
-                         [epilogue, ib](float acc, int im, int in) {
+                         [epilogue, ib](float acc, int64_t im, int64_t in) {
                              return epilogue(acc, ib, im, in);
                          });
   }
